refactor: 546 - move course-item to entity course
