+++ src/entities/course/ui/course-item/course-item.module.scss
